+++ common/db/src/main/java/module-info.java
@@ -33,6 +33,7 @@
 module is.codion.common.db {
 	requires transitive org.jspecify;
 	requires transitive java.sql;
+	requires org.slf4j;
 	requires transitive is.codion.common.core;
 
 	exports is.codion.common.db.connection;